+++ src/http/api/v1/index.ts
@@ -25,6 +25,7 @@ export const createHandler = (config: HttpApiConfig): express.Router => {
         return;
       }
 
+      console.error(err);
       config.logger.error({ msg: "Unexpected exception", err });
 
       Sentry.captureException(err);
