+++ src/http/api/v1/matches.ts
@@ -1,5 +1,6 @@
 import express, { Response, Request } from "express";
 import { z } from "zod";
+import { StaticPool } from "node-worker-threads-pool";
 
 const upload = multer();
 import multer from "multer";
@@ -10,10 +11,27 @@ import Calculator, {
   CalculatorOptions,
   parseOverrides,
 } from "../../../calculator/index.js";
+import { estimateMatches } from "../../../calculator/estimate.js";
 import { HttpApiConfig } from "../../app.js";
+import { potentialVoteSchema } from "../../../calculator/estimate.js";
+import { Round } from "../../../indexer/types.js";
+import {
+  CalculatorArgs,
+  CalculatorResult,
+} from "../../../calculator/worker.js";
+import { RoundContributionsCache } from "../../../calculator/roundContributionsCache.js";
 
 export const createHandler = (config: HttpApiConfig): express.Router => {
   const router = express.Router();
+
+  const calculatorWorkerPool = new StaticPool<
+    (msg: CalculatorArgs) => CalculatorResult
+  >({
+    size: 10,
+    task: "./dist/src/calculator/worker.js",
+  });
+
+  const roundContributionsCache = new RoundContributionsCache();
 
   function boolParam<T extends Record<string, unknown>>(
     object: T,
@@ -113,12 +131,16 @@ export const createHandler = (config: HttpApiConfig): express.Router => {
     await estimateMatchesHandler(req, res, 200);
   });
 
+  const estimateRequestBody = z.object({
+    potentialVotes: z.array(potentialVoteSchema),
+  });
+
   async function estimateMatchesHandler(
     req: Request,
     res: Response,
     okStatusCode: number
   ) {
-    const reqId = Math.random().toString(36).substring(7);
+    // const reqId = Math.random().toString(36).substring(7);
 
     const chainId = Number(req.params.chainId);
     const roundId = req.params.roundId;
@@ -130,27 +152,38 @@ export const createHandler = (config: HttpApiConfig): express.Router => {
       }));
 
     const chainConfig = config.chains.find((c) => c.id === chainId);
+
     if (chainConfig === undefined) {
       throw new ClientError(`Chain ${chainId} not configured`, 400);
     }
 
-    const calculatorOptions: CalculatorOptions = {
-      priceProvider: config.priceProvider,
-      dataProvider: config.dataProvider,
-      chainId: chainId,
-      roundId: roundId,
-      minimumAmountUSD: undefined,
-      matchingCapAmount: undefined,
-      overrides: {},
-      passportProvider: config.passportProvider,
-      chain: chainConfig,
-    };
+    const rounds = await config.dataProvider.loadFile<Round>(
+      "rounds",
+      `${chainId}/rounds.json`
+    );
 
-    const calculator = new Calculator(calculatorOptions);
-    const matches = await calculator.estimateMatching(potentialVotes);
+    const round = rounds.find((r: Round) => r.id === roundId);
+
+    if (round === undefined) {
+      throw new ClientError(`Round ${roundId} not found`, 400);
+    }
+
+    const matches = await estimateMatches({
+      round,
+      chain: chainConfig,
+      potentialVotes,
+      dataProvider: config.dataProvider,
+      priceProvider: config.priceProvider,
+      passportProvider: config.passportProvider,
+      roundCalculationConfig: {},
+      roundContributionsCache,
+      calculate: (args) => calculatorWorkerPool.exec(args),
+    });
+
     const responseBody = JSON.stringify(matches, (_key, value) =>
       typeof value === "bigint" ? value.toString() : (value as unknown)
     );
+
     res.setHeader("content-type", "application/json");
     res.status(okStatusCode);
     res.send(responseBody);
@@ -158,21 +191,3 @@ export const createHandler = (config: HttpApiConfig): express.Router => {
 
   return router;
 };
-
-const potentialVoteSchema = z.object({
-  projectId: z.string(),
-  roundId: z.string(),
-  applicationId: z.string(),
-  token: z.string(),
-  voter: z.string(),
-  grantAddress: z.string(),
-  amount: z.coerce.bigint(),
-});
-
-const potentialVotesSchema = z.array(potentialVoteSchema);
-const estimateRequestBody = z.object({
-  potentialVotes: potentialVotesSchema,
-});
-
-export type PotentialVotes = z.infer<typeof potentialVotesSchema>;
-export type PotentialVote = z.infer<typeof potentialVoteSchema>;
