+++ src/test/http/app.test.ts
@@ -7,7 +7,7 @@ import { createHttpApi } from "../../http/app.js";
 import { AugmentedResult, DataProvider } from "../../calculator/index.js";
 import { PriceProvider } from "../../prices/provider.js";
 import { Logger } from "pino";
-import { PotentialVotes } from "../../http/api/v1/matches.js";
+import { PotentialVote } from "../../calculator/estimate.js";
 import { Chain } from "../../config.js";
 import { constants } from "ethers";
 import {
@@ -279,62 +279,80 @@ describe("server", () => {
       test("should estimate matching with new votes for projects", async () => {
         const expectedResults = [
           {
-            applicationId: "application-id-1",
-            capOverflow: "0",
-            chainId: 1,
-            contributionsCount: "4",
+            original: {
+              totalReceived: "1500",
+              contributionsCount: "4",
+              sumOfSqrt: "70",
+              sumOfSqrtRemainder: "0",
+              capOverflow: "0",
+              matchedWithoutCap: "1360",
+              matched: "1360",
+            },
+            estimated: {
+              totalReceived: "1500",
+              contributionsCount: "4",
+              sumOfSqrt: "70",
+              sumOfSqrtRemainder: "0",
+              capOverflow: "0",
+              matchedWithoutCap: "1",
+              matched: "1",
+            },
             difference: "-1359",
             differenceInUSD: 0,
-            matched: "1",
-            matchedUSD: 0,
-            matchedWithoutCap: "1",
-            payoutAddress: "grant-address-1",
-            projectId: "project-id-1",
+            applicationId: "application-id-1",
             recipient: "grant-address-1",
-            roundId: "0x1234",
-            sumOfSqrt: "70",
-            sumOfSqrtRemainder: "0",
-            totalReceived: "1500",
           },
           {
-            applicationId: "application-id-2",
-            capOverflow: "0",
-            chainId: 1,
-            contributionsCount: "8",
+            original: {
+              totalReceived: "1000",
+              contributionsCount: "7",
+              sumOfSqrt: "80",
+              sumOfSqrtRemainder: "0",
+              capOverflow: "0",
+              matchedWithoutCap: "2160",
+              matched: "2160",
+            },
+            estimated: {
+              totalReceived: "1000",
+              contributionsCount: "8",
+              sumOfSqrt: "80",
+              sumOfSqrtRemainder: "0",
+              capOverflow: "0",
+              matchedWithoutCap: "1",
+              matched: "1",
+            },
             difference: "-2159",
             differenceInUSD: 0,
-            matched: "1",
-            matchedUSD: 0,
-            matchedWithoutCap: "1",
-            payoutAddress: "grant-address-2",
-            projectId: "project-id-2",
+            applicationId: "application-id-2",
             recipient: "grant-address-2",
-            roundId: "0x1234",
-            sumOfSqrt: "80",
-            sumOfSqrtRemainder: "0",
-            totalReceived: "1000",
           },
           {
-            applicationId: "application-id-3",
-            capOverflow: "0",
-            chainId: 1,
-            contributionsCount: "8",
+            original: {
+              totalReceived: "3400",
+              contributionsCount: "7",
+              sumOfSqrt: "140",
+              sumOfSqrtRemainder: "0",
+              capOverflow: "0",
+              matchedWithoutCap: "6480",
+              matched: "6480",
+            },
+            estimated: {
+              totalReceived: "10000003400",
+              contributionsCount: "8",
+              sumOfSqrt: "100140",
+              sumOfSqrtRemainder: "0",
+              capOverflow: "0",
+              matchedWithoutCap: "9996",
+              matched: "9996",
+            },
             difference: "3516",
             differenceInUSD: 0,
-            matched: "9996",
-            matchedUSD: 0,
-            matchedWithoutCap: "9996",
-            payoutAddress: "grant-address-3",
-            projectId: "project-id-3",
+            applicationId: "application-id-3",
             recipient: "grant-address-3",
-            roundId: "0x1234",
-            sumOfSqrt: "100140",
-            sumOfSqrtRemainder: "0",
-            totalReceived: "10000003400",
           },
         ];
 
-        const potentialVotes: PotentialVotes = [
+        const potentialVotes: PotentialVote[] = [
           {
             roundId: "round-id-1",
             applicationId: "application-id-3",
